Filter invalid transactions out of the payload and return them
Add a `rejected` metadata field to Block and related types. Now,
instead of failing if a block contains invalid transactions, the
engine will simply execute the transactions that it can and record
the rest in the `rejected` field.

In Espresso mode, the OP-node uses this to keep invalid transactions
in the _batch_, and deterministically derive a _block_ that applies
only the valid ones. This allows the derivation pipeline to check
that the sequencer has included all transactions required by the
Espresso Sequencer, rather than trusting the sequencer to exclude
transactions that were invalid (as it does in legacy mode).

Invalid transactions are represented as raw bytes, so this mechanism
makes it possible to even handle transactions submitted through
Espresso which don't unmarshal correctly, although this is not
implemented yet.